Fix null check in print_list

+++ 0x12-singly_linked_lists/0-print_list.c
@@ -27,7 +27,7 @@ size_t print_list(const list_t *h)
 	size_t size;
 
 	if (!h)
-		return (NULL);
+		return (size_t) (NULL);
 
 	size = 0;
 	while (h->next)
